+++ lib/components/IndexCard/IndexCard.tsx
@@ -15,14 +15,13 @@ import PushPinIcon from "@mui/icons-material/PushPin";
 import PushPinOutlinedIcon from "@mui/icons-material/PushPinOutlined";
 import VisibilityIcon from "@mui/icons-material/Visibility";
 import VisibilityOffIcon from "@mui/icons-material/VisibilityOff";
+import Masonry from "@mui/lab/Masonry";
 import Box from "@mui/material/Box";
 import Collapse from "@mui/material/Collapse";
 import Fade from "@mui/material/Fade";
 import FormControl from "@mui/material/FormControl";
 import Grid from "@mui/material/Grid";
 import IconButton from "@mui/material/IconButton";
-import ImageList from "@mui/material/ImageList";
-import ImageListItem from "@mui/material/ImageListItem";
 import InputLabel from "@mui/material/InputLabel";
 import Link from "@mui/material/Link";
 import Paper from "@mui/material/Paper";
@@ -127,8 +126,6 @@ export const IndexCard: React.FC<
     indexCard: IIndexCard;
     parentIndexCard?: IIndexCard;
     allCards: Array<IIndexCard>;
-    reactDndIndex: number;
-    reactDndType: string;
     canMove: boolean;
     isGM: boolean;
     indexCardHiddenRecord?: Record<string, boolean>;
@@ -137,7 +134,7 @@ export const IndexCard: React.FC<
     onMoveTo?(idOfIndexCardToMove: string, idOfIndexCardToMoveTo: string): void;
     onMoveOut?(idOfIndexCardToMove: string): void;
     onRoll(diceRollResult: IDiceRollResult): void;
-    onMove(dragIndex: number, hoverIndex: number): void;
+    onMove(dragIndex: string, hoverIndex: string): void;
     onRemove(): void;
     onDuplicate(): void;
     onTogglePrivate?(): void;
@@ -148,7 +145,6 @@ export const IndexCard: React.FC<
   const { t } = useTranslate();
 
   const [hover, setHover] = useState(false);
-
   const $menu = useRef(null);
   const indexCardManager = useIndexCard({
     indexCard: props.indexCard,
@@ -207,10 +203,14 @@ export const IndexCard: React.FC<
                 <BetterDnd
                   direction="horizontal"
                   key={indexCardManager.state.indexCard.id}
-                  index={props.reactDndIndex}
-                  type={props.reactDndType}
-                  onMove={(dragIndex, hoverIndex) => {
-                    props.onMove(dragIndex, hoverIndex);
+                  id={indexCardManager.state.indexCard.id}
+                  type={
+                    hasSubCards
+                      ? `${DragAndDropTypes.SceneIndexCardsSubCards}-${indexCardManager.state.indexCard.id}`
+                      : DragAndDropTypes.SceneIndexCards
+                  }
+                  onMove={(dragId, hoverId) => {
+                    props.onMove(dragId, hoverId);
                   }}
                   render={(dndRenderProps) => {
                     return (
@@ -532,70 +532,47 @@ export const IndexCard: React.FC<
   function renderSubCards() {
     return (
       <Box px="1rem" py="1rem">
-        <ImageList
-          variant={"standard"}
-          cols={numberOfColumnsForSubCards}
-          gap={16}
-          className={css({
-            // padding: "1rem 1rem", // for boxShadow padding
-            // margin: "-1rem -1rem", // for boxShadow padding
+        <Masonry columns={numberOfColumnsForSubCards}>
+          {indexCardManager.state.indexCard.subCards?.map((subCard) => {
+            return (
+              <Box key={subCard.id}>
+                <IndexCard
+                  indexCard={subCard}
+                  isGM={props.isGM}
+                  parentIndexCard={indexCardManager.state.indexCard}
+                  allCards={props.allCards}
+                  id={`index-card-${subCard.id}`}
+                  canMove={true}
+                  onPoolClick={props.onPoolClick}
+                  onRoll={props.onRoll}
+                  indexCardHiddenRecord={props.indexCardHiddenRecord}
+                  onToggleVisibility={props.onToggleVisibility}
+                  onMoveTo={(idOfIndexCardToMove, idOfIndexCardToMoveTo) => {
+                    props.onMoveTo?.(
+                      idOfIndexCardToMove,
+                      idOfIndexCardToMoveTo
+                    );
+                  }}
+                  onMoveOut={(idOfIndexCardToMove) => {
+                    props.onMoveOut?.(idOfIndexCardToMove);
+                  }}
+                  onMove={(dragId, hoverId) => {
+                    indexCardManager.actions.moveIndexCard(dragId, hoverId);
+                  }}
+                  onChange={(newIndexCard) => {
+                    indexCardManager.actions.updateIndexCard(newIndexCard);
+                  }}
+                  onDuplicate={() => {
+                    indexCardManager.actions.duplicateIndexCard(subCard);
+                  }}
+                  onRemove={() => {
+                    indexCardManager.actions.removeIndexCard(subCard.id);
+                  }}
+                />
+              </Box>
+            );
           })}
-        >
-          {indexCardManager.state.indexCard.subCards?.map(
-            (subCard, subCardIndex) => {
-              return (
-                <ImageListItem
-                  key={subCard.id}
-                  cols={1}
-                  className={css({
-                    width: "100%",
-                    paddingTop: ".25rem",
-                    paddingBottom: ".25rem",
-                  })}
-                >
-                  <IndexCard
-                    indexCard={subCard}
-                    isGM={props.isGM}
-                    parentIndexCard={indexCardManager.state.indexCard}
-                    allCards={props.allCards}
-                    id={`index-card-${subCard.id}`}
-                    reactDndType={`${DragAndDropTypes.SceneIndexCardsSubCards}.${indexCardManager.state.indexCard.id}`}
-                    canMove={true}
-                    reactDndIndex={subCardIndex}
-                    onPoolClick={props.onPoolClick}
-                    onRoll={props.onRoll}
-                    indexCardHiddenRecord={props.indexCardHiddenRecord}
-                    onToggleVisibility={props.onToggleVisibility}
-                    onMoveTo={(idOfIndexCardToMove, idOfIndexCardToMoveTo) => {
-                      props.onMoveTo?.(
-                        idOfIndexCardToMove,
-                        idOfIndexCardToMoveTo
-                      );
-                    }}
-                    onMoveOut={(idOfIndexCardToMove) => {
-                      props.onMoveOut?.(idOfIndexCardToMove);
-                    }}
-                    onMove={(dragIndex, hoverIndex) => {
-                      indexCardManager.actions.moveIndexCard(
-                        dragIndex,
-                        hoverIndex
-                      );
-                    }}
-                    onChange={(newIndexCard) => {
-                      indexCardManager.actions.updateIndexCard(newIndexCard);
-                    }}
-                    onDuplicate={() => {
-                      indexCardManager.actions.duplicateIndexCard(subCard);
-                    }}
-                    onRemove={() => {
-                      indexCardManager.actions.removeIndexCard(subCard.id);
-                    }}
-                  />
-                </ImageListItem>
-              );
-            }
-          )}
-        </ImageList>
+        </Masonry>
       </Box>
     );
   }
@@ -607,18 +584,15 @@ export const IndexCard: React.FC<
 
     return (
       <Box px="1rem" py=".5rem">
-        {indexCardManager.state.indexCard.blocks.map((block, blockIndex) => {
+        {indexCardManager.state.indexCard.blocks.map((block) => {
           return (
             <Box key={block.id} mb=".5rem">
               <BetterDnd
                 direction="vertical"
-                index={blockIndex}
+                id={block.id}
                 type={`${DragAndDropTypes.SceneIndexCardsBlocks}.${indexCardManager.state.indexCard.id}`}
-                onMove={(dragIndex, hoverIndex) => {
-                  indexCardManager.actions.moveIndexCardBlock(
-                    dragIndex,
-                    hoverIndex
-                  );
+                onMove={(dragId, hoverId) => {
+                  indexCardManager.actions.moveIndexCardBlock(dragId, hoverId);
                 }}
                 render={(dndRenderProps) => {
                   return (