+++ lib/routes/Character/components/BetterDnD/BetterDnd.tsx
@@ -1,20 +1,17 @@
 import { css, cx } from "@emotion/css";
 import { useTheme } from "@mui/material/styles";
 import React, { useRef, useState } from "react";
-import {
-  ConnectDragSource,
-  DropTargetMonitor,
-  useDrag,
-  useDrop,
-  XYCoord,
-} from "react-dnd";
+import { ConnectDragSource, useDrag, useDrop } from "react-dnd";
 
+type IItem = {
+  id: string;
+};
 export const BetterDnd: React.FC<{
-  index: number;
   /**
    * Unique key to know where blocks can be dropped
    */
   type: string;
+  id: string;
   className?: string;
   render(renderProps: {
     drag: ConnectDragSource;
@@ -24,7 +21,7 @@ export const BetterDnd: React.FC<{
   onDrag?(): void;
   onDrop?(): void;
   direction: "horizontal" | "vertical";
-  onMove?(dragIndex: number, hoverIndex: number): void;
+  onMove?(dragId: string, hoverId: string): void;
 }> = (props) => {
   const theme = useTheme();
   const ref = useRef<HTMLDivElement>(null);
@@ -34,84 +31,32 @@ export const BetterDnd: React.FC<{
     item: () => {
       props.onDrag?.();
       return {
-        id: props.index,
+        id: props.id,
       };
     },
-
     collect: (monitor) => ({
       isDragging: !!monitor.isDragging(),
     }),
   });
 
-  const [{}, drop] = useDrop({
+  const [, drop] = useDrop<IItem>({
     accept: props.type,
     drop: () => {
       props.onDrop?.();
     },
-    hover(item: any, monitor: DropTargetMonitor) {
+    hover(item: IItem) {
       if (!ref.current) {
         return;
       }
-      const dragIndex = item.index;
-      const hoverIndex = props.index;
+      const dragId = item.id;
+      const hoverId = props.id;
 
       // Don't replace items with themselves
-      if (dragIndex === hoverIndex) {
+      if (dragId === hoverId) {
         return;
       }
 
-      // Determine rectangle on screen
-      const hoverBoundingRect = ref.current?.getBoundingClientRect();
-
-      // Get vertical middle
-      const hoverMiddleY =
-        (hoverBoundingRect.bottom - hoverBoundingRect.top) / 2;
-      const hoverMiddleX =
-        (hoverBoundingRect.right - hoverBoundingRect.left) / 2;
-
-      // Determine mouse position
-      const clientOffset = monitor.getClientOffset();
-
-      // Get pixels to the top
-      const hoverClientY = (clientOffset as XYCoord).y - hoverBoundingRect.top;
-      const hoverClientX = (clientOffset as XYCoord).x - hoverBoundingRect.left;
-
-      // Only perform the move when the mouse has crossed half of the items height
-      // When dragging downwards, only move when the cursor is below 50%
-      // When dragging upwards, only move when the cursor is above 50%
-      // Dragging downwards
-
-      if (props.direction === "vertical") {
-        if (dragIndex < hoverIndex && hoverClientY < hoverMiddleY) {
-          return;
-        }
-
-        // Dragging upwards
-        if (dragIndex > hoverIndex && hoverClientY > hoverMiddleY) {
-          return;
-        }
-      }
-      if (props.direction === "horizontal") {
-        // if (dragIndex < hoverIndex && hoverClientY < hoverMiddleY) {
-        if (dragIndex < hoverIndex && hoverClientX < hoverMiddleX) {
-          return;
-        }
-
-        // Dragging rightwards
-        // if (dragIndex > hoverIndex && hoverClientY > hoverMiddleY) {
-        if (dragIndex > hoverIndex && hoverClientX > hoverMiddleX) {
-          return;
-        }
-      }
-
-      // Time to actually perform the action
-      props.onMove?.(dragIndex, hoverIndex);
-
-      // Note: we're mutating the monitor item here!
-      // Generally it's better to avoid mutations,
-      // but it's good here for the sake of performance
-      // to avoid expensive index searches.
-      item.index = hoverIndex;
+      props.onMove?.(dragId, hoverId);
     },
     collect: (monitor) => ({
       isOver: !!monitor.isOver(),